libmantalite
