+++ lib/common.js
@@ -22,6 +22,7 @@ var restifyErrors = require('restify-errors');
 var VError = require('verror');
 
 var CheckStream = require('./check_stream');
+var libmantalite = require('./libmantalite');
 require('./errors');
 var muskieUtils = require('./utils');
 var sharkClient = require('./shark_client');
@@ -31,8 +32,6 @@ var utils = require('./utils');
 
 var clone = utils.shallowCopy;
 var sprintf = util.format;
-
-var ANONYMOUS_USER = 'anonymous';
 
 var CORS_RES_HDRS = [
     'access-control-allow-headers',
@@ -791,36 +790,6 @@ function getSharkInfo(req, hostname) {
     return (sharks[0]);
 }
 
-/**
- * Given an resource like `/mark/stor/my%resume.pdf`, this API will convert to
- * `/mark/stor/my resume.pdf`; if you optionally pass in an `account` object as
- * returned from mahi.js (in this lib), the path will be converted to
- * `/d3d38d1a-e26c-11e2-9a8a-475696be42f1/stor/my resume.pdf`, which is the
- * correct normal form everwhere in the system.
- *
- * Note that because `decodeURIComponent` may throw, this function takes a
- * callback, even though it is not blocking.
- */
-function normalizeMantaPath(opts, cb) {
-    assert.object(opts, 'options');
-    assert.string(opts.path, 'options.path');
-    assert.optionalObject(opts.account, 'options.account');
-
-    var err = null;
-    var key;
-
-    try {
-        key = opts.path.split('/').map(decodeURIComponent).join('/');
-
-        if (opts.account)
-            key = key.replace(PATH_LOGIN_RE, '/' + opts.account.uuid + '/');
-
-    } catch (e) {
-        err = e;
-    }
-
-    cb(err, key);
-}
 
 ///--- Exports
 
@@ -830,8 +799,6 @@ module.exports = {
     DEF_MAX_COPIES: DEF_MAX_COPIES,
     DEF_NUM_COPIES: DEF_NUM_COPIES,
     ZERO_BYTE_MD5: ZERO_BYTE_MD5,
-
-    ANONYMOUS_USER: ANONYMOUS_USER,
 
     CORS_RES_HDRS: CORS_RES_HDRS,
 
@@ -974,7 +941,7 @@ module.exports = {
                 path: req.path()
             };
 
-            normalizeMantaPath(_opts, function (err, p) {
+            libmantalite.normalizeMantaPath(_opts, function (err, p) {
                 if (err) {
                     req.log.debug({
                         url: req.path(),
@@ -999,6 +966,5 @@ module.exports = {
     findSharks: findSharks,
     startSharkStreams: startSharkStreams,
     sharkStreams: sharkStreams,
-    streamFromSharks: streamFromSharks,
-    normalizeMantaPath: normalizeMantaPath
+    streamFromSharks: streamFromSharks
 };
