+++ lib/uploads/create.js
@@ -26,6 +26,7 @@ require('ajv-keywords')(ajv, ['prohibited']);
 
 var auth = require('../auth');
 var common = require('../common');
+var libmantalite = require('../libmantalite');
 var obj = require('../obj');
 var uploadsCommon = require('./common');
 var muskieUtils = require('../utils');
@@ -375,7 +376,7 @@ function loadTargetObjectOwner(req, res, next) {
 
 
     // We don't support mpu for users yet.
-    var user = common.ANONYMOUS_USER;
+    var user = libmantalite.ANONYMOUS_USER;
     var fallback = true;
     req.mahi.getUser(user, account, fallback, function (err, owner) {
         if (err) {
